pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "com.creatumlibre.www"
version = "0.1.0"
description = "Ein kreatives Grafik-Projekt mit Python"
authors = [{name = "Martin Stottmeister", email = "marty@umbrella-brothers.de"}]
dependencies = [
    "opencv-python",  # Bildverarbeitung
    "cairocffi",      # Cairo-Bindings für Grafik
    "pillow",         # Rastergrafik-Handling
    "numpy",          # Mathematische Operationen
    "pycairo",        # Alternative Cairo-Bindings
    "matplotlib",     # Plots & Visualisierung
    "scipy",          # Wissenschaftliche Berechnungen
    "wand"            # ImageMagick-Integration
]
requires-python = ">=3.13"

[tool.setuptools.packages.find]
where = ["src"]
exclude =["test*"]

[tool.isort]
profile = "black"

[project.scripts]
creatumlibre = "main:main"
